Chaned sample count type to long unsigned, instead of just unsigned. This means that in theory 64bit builds could work with 24 hours of sound in FS=44100. that is

+++ signal-audiofile.h
@@ -242,7 +242,7 @@ public:
 
     virtual pBuffer read( unsigned firstSample, unsigned numberOfSamples );
     virtual unsigned sample_rate();
-    virtual unsigned number_of_samples();
+    virtual long unsigned number_of_samples();
 
     pBuffer getChunkBehind() { return _waveform; }
     void setChunk( pBuffer chunk ) { _waveform = chunk; }
